+++ test/411-queue-translation-hit.cc
@@ -1,13 +1,13 @@
 #include "catch.hpp"
 #include "mocks.hpp"
-#include "cache.h"
+#include "channel.h"
 #include "champsim_constants.h"
 
 SCENARIO("Cache queues issue translations in WQ") {
   GIVEN("A write queue with one item") {
     constexpr uint64_t hit_latency = 1;
     do_nothing_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
@@ -52,7 +52,7 @@ SCENARIO("Cache queues issue translations in RQ") {
   GIVEN("A read queue with one item") {
     constexpr uint64_t hit_latency = 1;
     do_nothing_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
@@ -97,7 +97,7 @@ SCENARIO("Cache queues issue translations in PQ") {
   GIVEN("A prefetch queue with one item") {
     constexpr uint64_t hit_latency = 1;
     do_nothing_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
@@ -141,7 +141,7 @@ SCENARIO("Translations in the WQ work even if the addresses happen to be the sam
   GIVEN("A write queue with one item") {
     constexpr uint64_t hit_latency = 1;
     release_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
@@ -187,7 +187,7 @@ SCENARIO("Translations in the RQ work even if the addresses happen to be the sam
   GIVEN("A read queue with one item") {
     constexpr uint64_t hit_latency = 1;
     release_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
@@ -233,7 +233,7 @@ SCENARIO("Translations in the PQ work even if the addresses happen to be the sam
   GIVEN("A prefetch queue with one item") {
     constexpr uint64_t hit_latency = 1;
     release_MRC mock_ll;
-    CACHE::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
+    champsim::TranslatingQueues uut{1, 32, 32, 32, 0, hit_latency, LOG2_BLOCK_SIZE, false};
     uut.lower_level = &mock_ll;
 
     // Turn off warmup
